+++ manager.mjs
@@ -1,11 +1,9 @@
 //--------------------------------------------------------
 //-- Manager
 //--------------------------------------------------------
-'use strict';
-
-const path        = require('path');
-const fss         = require('@absolunet/fss');      // eslint-disable-line node/no-unpublished-require
-const { manager } = require('@absolunet/manager');  // eslint-disable-line node/no-unpublished-require
+import { fileURLToPath, URL } from 'node:url';
+import fss                    from '@absolunet/fss';      // eslint-disable-line node/no-unpublished-import
+import { manager }            from '@absolunet/manager';  // eslint-disable-line node/no-unpublished-import,import/no-unresolved
 
 manager.init({
 	repositoryType: 'multi-package',
@@ -17,9 +15,9 @@ manager.init({
 			postRun: ({ terminal }) => {
 				terminal.print(`Sync Node.js version in all package.json`).spacer();
 
-				const nodeVersion = fss.readJson(path.join(__dirname, 'package.json')).engines.node;
+				const nodeVersion = fss.readJson(fileURLToPath(new URL('package.json', import.meta.url))).engines.node;
 
-				fss.scandir(path.join(__dirname, 'packages'), 'dir', { fullPath: true }).forEach((root) => {
+				fss.scandir(fileURLToPath(new URL('packages', import.meta.url)), 'dir', { fullPath: true }).forEach((root) => {
 					const packageFile = `${root}/package.json`;
 					const packageData = fss.readJson(packageFile);
 					packageData.engines.node = nodeVersion;