+++ packages/test/test/generic/index.test.js
@@ -2,6 +2,5 @@
 //-- Tester
 //--------------------------------------------------------
 import { tester } from '@absolunet/tester';
-import fileMatrix from '../../../../test/generic/matrix/file-matrix';
 
-tester.genericRepositoryTests({ fileMatrix });
+tester.genericRepositoryTests();