+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Remove `@magento/eslint-config` because it is not maintained and is incompatible with recent version of ESLint
+- Manager and tester updates